fixing hirarki, and some bug in setup to do list (set to do data from string to array)

+++ js/commands/settings/todo.js
@@ -1,20 +1,20 @@
 const { MessageEmbed } = require('discord.js')
 const { getTodoDB, updateTodoDB } = require('../../db')
 
-const ynString = [ 'ya', 'tidak' ]
+const defaultTemp = {
+	default: '🔸',
+	onGoing: '🔹',
+	completed: '✅',
+	uncompleted: '📛'
+}
+/*
 
-let setAuthor
 
-const filterAuthor = m => setAuthor === m.author.id
-const filterCondition = m => filterAuthor(m) && ynString.some(b => b === m.content.toLowerCase())
-const filterOneLine = m => filterAuthor(m) && m.content.split('\n').length === 1
-const filterNumber = m => filterAuthor(m) && eval(m.content) !== isNaN
-const filterNumbers = m => filterAuthor(m) && m.content.split(',').every(num => eval(num) !== isNaN)
 
+*/
 module.exports = {
 	name: 'todo',
 	execute(msg, args) {
-		setAuthor = msg.author.id
 		const userNickname = msg.guild.members.cache.get(msg.author.id).nickname
 		const settingsDesc = 'Tekan reaction di bawah untuk mengatur to do list anda:\n\n➕ = `tambah list`\n🗑️ = `hapus beberapa list`\n📝 = `mengedit list`\n📦 = `template todolist`\n✅ = `selesai`'
 		const settingsEmbed = new MessageEmbed()
@@ -40,8 +40,7 @@ module.exports = {
 							m.delete()
 							if (!list) return addTodoList(msg)
 
-							const todoList = list.map(item => item.join(' pada '))
-							addTodoList(msg, todoList)
+							addTodoList(msg, list)
 						})
 						return
 					case '🗑️':
@@ -51,6 +50,7 @@ module.exports = {
 							removeTodoList(msg, list)
 						})
 						return
+
 					case '📝': 
 						getTodoDB(msg.author.id).then(list => {
 							m.delete()
@@ -58,11 +58,23 @@ module.exports = {
 							editTodoList(msg, list)
 						})
 						return
+
 					case '📦':
+						m.delete()
 						msg.channel.send('belum ada template diatur untuk server ini')
 						return
+
 					case '✅':
-						console.log('done')
+						getTodoDB(msg.author.id).then(list => {							
+							m.delete()
+							const todoEmbed = new MessageEmbed()
+								.setColor('#347C7C')
+								.setAuthor(msg.author.tag, msg.author.displayAvatarURL())
+								.setTitle(`DAILY TO DO LIST`)
+								.setDescription(`▫️ \`${list.map(item => item.join(' pada ')).join('\`\n▫️ \`')}\``)
+								.setFooter('gunakan p!set todo untuk mengedit list')
+							msg.channel.send(todoEmbed)
+						})
 						return
 				}
 			})
@@ -74,13 +86,26 @@ async function addTodoList(msg, todoList = []) {
 	
 	const todoArray = todoList
 
-	const qTxt1 = `**<@${msg.author.id}>, silahkan Masukkan Todo List Kamu hari ini**\n\nGunakan:\n\`enter\` untuk memasukkan lebih dari 1 list sekaligus\n\`-at\` untuk mengisi Waktu Pengerjaan\n\nContoh:\n\`Solat subuh & olahraga -at 04.00 - 05.00\nSiap-siap Zoom -at 05.00 - 07.00\nSekolah Online -at 07.00 - Selesai\``
+	const ynString = [ 'ya', 'tidak' ]
+
+	const filterAuthor = m => msg.author.id === m.author.id
+	const filterExit = m => m.content.toLowerCase() === 'exit'
+	const filterCondition = m => filterAuthor(m) && ( filterExit(m)  || ynString.some(b => b === m.content.toLowerCase()) )
+	const filterOneLine = m => filterAuthor(m) && ( filterExit(m)  || m.content.split('\n').length === 1 )
+	const filterNumber = m => filterAuthor(m) && ( filterExit(m)  || eval(m.content) !== isNaN )
+	const filterNumbers = m => filterAuthor(m) && ( filterExit(m)  || m.content.split(',').every(num => eval(num) !== isNaN) )
+
+
+	const qTxt1 = `**<@${msg.author.id}>, silahkan Masukkan Todo List Kamu hari ini**. Gunakan \`enter\` (baris baru) untuk memasukkan lebih dari 1 list sekaligus\n\`\`\`\nContoh:\n**(04.00-07.00)** Rutinitas pagi\n**(07.00-11.00)** Sekolah Online\nIstirahat siang\nEkskul *Tata Boga*\n\`\`\``
 
 	const input1 = await awaitSingleMessage(msg, filterAuthor, qTxt1)
 	const todoInput = input1.split('\n')
 	const newTodoArray = [ ...todoArray, ...todoInput]
 
-	const qTxt2 = `\`\`\`\nTO DO LIST HARI INI:\n${newTodoArray.join('\n')}\n\`\`\`\n Apakah Kamu ingin menambah to do list lagi? **(Ya/Tidak)**`
+	const qTxt2 = new MessageEmbed()
+		.setColor('#347C7C')
+		.setAuthor("TODO LIST HARI INI:", msg.author.displayAvatarURL())
+		.setDescription(`\n▫️ ${newTodoArray.join('\n▫️ ')}\n\n\`\`\`json\nApakah Kamu ingin menambah to do list lagi? "(Ketik: Ya/Tidak)"\n\`\`\``)
 
 	const input2 = await awaitSingleMessage(msg, filterCondition, qTxt2)
 	const isAddAgain = input2.toLowerCase() === 'ya'
@@ -90,27 +115,35 @@ async function addTodoList(msg, todoList = []) {
 		return
 	}
 
-	const todoData = newTodoArray.map(item =>  item.split(' -at '))
-	const todoEmbed = new MessageEmbed()
+	let todoEmbed = new MessageEmbed()
 		.setColor('#347C7C')
 		.setAuthor(msg.author.tag, msg.author.displayAvatarURL())
 		.setTitle(`DAILY TO DO LIST`)
-		.setDescription(`▫️ \`${newTodoArray.map(item => item.replace(' -at ', ' pada ')).join('\`\n▫️ \`')}\``)
-		.setFooter('gunakan p!set todo untuk mengedit list')
+		.setDescription(`▫️ ${newTodoArray.join('\n▫️ ')}`)
+		.setFooter('gunakan `,p set todo` untuk mengedit list')
 	msg.channel.send(todoEmbed)
-	updateTodoDB(msg.author.id, todoData)
+	updateTodoDB(msg.author.id, newTodoArray)
 }
 
 async function removeTodoList(msg, todoList) {
 	let newTodoData = todoList
 
-	const todoString = todoList.map((item, index) => `${index + 1}. ${item[0]} **(${item[1]})**`)
+	if (newTodoData.length < 1) return msg.channel.send('Todo List kamu hari ini kosong')
+
+	const ynString = [ 'ya', 'tidak' ]
+
+	const filterAuthor = m => msg.author.id === m.author.id
+	const filterExit = m => m.content.toLowerCase() === 'exit'
+	const filterCondition = m => filterAuthor(m) && ( filterExit(m)  || ynString.some(b => b === m.content.toLowerCase()) )
+	const filterOneLine = m => filterAuthor(m) && ( filterExit(m)  || m.content.split('\n').length === 1 )
+	const filterNumber = m => filterAuthor(m) && ( filterExit(m)  || eval(m.content) !== isNaN )
+	const filterNumbers = m => filterAuthor(m) && ( filterExit(m)  || m.content.split(',').every(num => eval(num) !== isNaN) )
+
 	const todoEmbed = new MessageEmbed()
 		.setColor('#347C7C')
-		.setAuthor(msg.author.tag, msg.author.displayAvatarURL())
-		.setTitle(`DAILY TO DO LIST`)
-		.setDescription(todoString.join('\n'))
-		.setFooter('ketik nomor list yang ingin dihapus')
+		.setAuthor(`DAILY TO DO LIST`, msg.author.displayAvatarURL())
+		.setDescription(`\`\`\`json\n"Masukkan Nomor List yang ingin Dihapus:"\`\`\`\n${newTodoData.map((item, index) => `**${index + 1}.** ${item}`).join('\n')}\n`)
+		.setFooter('ketik `exit` untuk membatalkan proses')
 
 	const input = await awaitSingleMessage(msg, filterNumbers, todoEmbed)
 	const inputArray = input.split(',').map(num => eval(num) - 1).sort((a, b) => b - a)
@@ -121,13 +154,12 @@ async function removeTodoList(msg, todoList) {
 	const isDelete = input2.toLowerCase() === 'ya'
 
 	if (!isDelete) {
-		msg.channel.send('**Selesai!** Gunakan `p!todo` untuk melihat list dan gunakan `p!set todo` untuk kembali mengatur list')		
-		updateTodoDB(msg.author.id, newTodoData)
+		removeTodoList(msg, newTodoArray)
 		return 
 	}
 
 	for (const num of inputArray) {
-		if (num > newTodoData.length - 1) {
+		if (isNaN(num) || num > newTodoData.length - 1) {
 			msg.channel.send('masukkan nomor to do list dengan benar')
 			removeTodoList(msg, newTodoData)
 			return
@@ -135,7 +167,10 @@ async function removeTodoList(msg, todoList) {
 		newTodoData.splice(num, 1)
 	}
 
-	const qTxt3 = `\`\`\`\nTO DO LIST HARI INI:\n${newTodoData.join('\n')}\n\`\`\`\n Apakah kamu ingin menghapus to do list lain? **(Ya/Tidak)**`
+	const qTxt3 = new MessageEmbed()
+		.setColor('#347C7C')
+		.setAuthor("TODO LIST HARI INI:", msg.author.displayAvatarURL())
+		.setDescription(`\n▫️ ${newTodoData.join('\n▫️ ')}\n\n\`\`\`json\nApakah Kamu ingin menghapus to do list lagi? "(Ketik: Ya/Tidak)"\n\`\`\``)
 
 	const input3 = await awaitSingleMessage(msg, filterCondition, qTxt3)
 	const isAddAgain = input3.toLowerCase() === 'ya'
@@ -145,20 +180,27 @@ async function removeTodoList(msg, todoList) {
 		return
 	}
 
-	msg.channel.send('**Selesai!** Gunakan `p!todo` untuk melihat list dan gunakan `p!set todo` untuk kembali mengatur list')
+	msg.channel.send('**Selesai!** Gunakan `,p todo` untuk melihat list dan gunakan `p!set todo` untuk kembali mengatur list')
 	updateTodoDB(msg.author.id, newTodoData)
 }
 
 async function editTodoList(msg, todoList) {
 	let newTodoData = todoList
 
-	const todoString = todoList.map((item, index) => `${index + 1}. ${item[0]} **(${item[1]})**`)
+	const ynString = [ 'ya', 'tidak' ]
+
+	const filterAuthor = m => msg.author.id === m.author.id
+	const filterExit = m => m.content.toLowerCase() === 'exit'
+	const filterCondition = m => filterAuthor(m) && ( filterExit(m)  || ynString.some(b => b === m.content.toLowerCase()) )
+	const filterOneLine = m => filterAuthor(m) && ( filterExit(m)  || m.content.split('\n').length === 1 )
+	const filterNumber = m => filterAuthor(m) && ( filterExit(m)  || eval(m.content) !== NaN )
+	const filterNumbers = m => filterAuthor(m) && ( filterExit(m)  || m.content.split(',').every(num => eval(num) !== NaN) )
+
 	const todoEmbed = new MessageEmbed()
 		.setColor('#347C7C')
-		.setAuthor(msg.author.tag, msg.author.displayAvatarURL())
-		.setTitle(`DAILY TO DO LIST`)
-		.setDescription(todoString.join('\n'))
-		.setFooter('ketik nomor list yang ingin diedit')
+		.setAuthor(`DAILY TO DO LIST`, msg.author.displayAvatarURL())
+		.setDescription(`\`\`\`json\n"Masukkan Nomor List yang ingin Diedit:"\`\`\`\n${newTodoData.map((item, index) => `**${index + 1}.** ${item}`).join('\n')}\n`)
+		.setFooter('ketik `exit` untuk membatalkan proses')
 
 	const inputNum = await awaitSingleMessage(msg, filterNumber, todoEmbed)
 	const itemNum = eval(inputNum) - 1
@@ -169,14 +211,16 @@ async function editTodoList(msg, todoList) {
 		return
 	}
 
-	const qTxt2 = `Masukkan to do baru untuk nomor ${inputNum}\nGunakan \`-at\` untuk mengatur waktu`
+	const qTxt2 = `Masukkan to do baru untuk nomor ${inputNum}:\n*${newTodoData[itemNum]}*`
 
 	const inputItem = await awaitSingleMessage(msg, filterOneLine, qTxt2)
-	const newItem = inputItem.split(' -at ')
 
-	newTodoData[itemNum] = newItem
+	newTodoData[itemNum] = inputItem
 
-	const qTxt3 = `\`\`\`\nTO DO LIST HARI INI:\n${newTodoData.join('\n')}\n\`\`\`\n Apakah kamu ingin mengedit to do list lain? **(Ya/Tidak)**`
+	const qTxt3 = new MessageEmbed()
+		.setColor('#347C7C')
+		.setAuthor("TODO LIST HARI INI:", msg.author.displayAvatarURL())
+		.setDescription(`\n▫️ ${newTodoData.join('\n▫️ ')}\n\n\`\`\`json\nApakah Kamu ingin mengedit to do list lagi? "(Ketik: Ya/Tidak)"\n\`\`\``)
 
 	const input3 = await awaitSingleMessage(msg, filterCondition, qTxt3)
 	const isAddAgain = input3.toLowerCase() === 'ya'
@@ -194,8 +238,9 @@ async function awaitSingleMessage(msg, filter, questionTxt) {
 	const questionMsg = await msg.channel.send(questionTxt)
 	const input = await msg.channel.awaitMessages(filter, { max: 1 }).then(collected => Promise.resolve(collected.first()))
 
-	questionMsg.delete()
-	input.delete()
 
+	questionMsg.delete()
+	if (input.content.toLowerCase() === 'exit') return msg.channel.send('**Proses Dihentikan**')
+	input.delete()
 	return Promise.resolve(input.content)
 }
